% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_overview.R
\name{add_sub_month}
\alias{add_sub_month}
\title{Add or subtract a month.}
\usage{
add_sub_month(selected_date, add = TRUE)
}
\arguments{
\item{selected_date}{date to perform operation.}

\item{add}{True to add a month false to subtract a month}
}
\value{
date type.
}
\description{
Add or subtract a month.
}
\examples{
add_sub_month("2020-05-24", FALSE)

}
