% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_overview.R
\name{best_category}
\alias{best_category}
\title{<table> of the best category of product, customer, region, sales team}
\usage{
best_category(df, arrange_by = "sum")
}
\arguments{
\item{df}{data frame or a reactive data frame.}

\item{arrange_by}{sort the table using what variable, default is 'sum'.}
}
\value{
a react table output.
}
\description{
<table> of the best category of product, customer, region, sales team
}
\examples{
best_category(r_df())

}
