% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_general.R
\name{clean_text}
\alias{clean_text}
\title{<character> replace underscore with space and convert string to title.}
\usage{
clean_text(string)
}
\arguments{
\item{string}{a string to clean.}
}
\value{
a character vector.
}
\description{
<character> replace underscore with space and convert string to title.
}
\examples{
clean_text("customer_name")

}
