% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_general.R
\name{count_chr}
\alias{count_chr}
\title{<data frame> count the number of observation in each character category.}
\usage{
count_chr(df, count_var, add_percent = TRUE)
}
\arguments{
\item{df}{a data frame.}

\item{count_var}{the character variable to count.}

\item{add_percent}{whether to add each category proportion of the total column.}
}
\value{
a tibble
}
\description{
<data frame> count the number of observation in each character category.
}
\examples{
 count_chr(data, "sales_team")

}
