% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_product.R
\name{count_chr_plt}
\alias{count_chr_plt}
\title{<plot> number of transactions}
\usage{
count_chr_plt(df, gp_var, top = TRUE, n = 10)
}
\arguments{
\item{df}{data frame or a reactive data frame.}

\item{gp_var}{character variable with unique values to count.}

\item{top}{whether to count the top n or bottom n.}

\item{n}{number of observations to return.}
}
\value{
a echarts4r bar chart.
}
\description{
<plot> number of transactions
}
\examples{
count_chr_plt(r_df(), "product_name")

}
