% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_customer.R
\name{customer_purchase_freq}
\alias{customer_purchase_freq}
\title{<table> showing the number of time each customer made a purchase.}
\usage{
customer_purchase_freq(df, page_size = 11)
}
\arguments{
\item{df}{data frame of a reactive data frame.}

\item{page_size}{the initial number of rows to display default 11.}
}
\value{
a react table output.
}
\description{
<table> showing the number of time each customer made a purchase.
}
\examples{
customer_purchase_freq(r_df())

}
