% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_customer.R
\name{customer_sales_profit_summary}
\alias{customer_sales_profit_summary}
\title{<table> showing the summary of sales, cost & profit from each customer.}
\usage{
customer_sales_profit_summary(df, s_fun)
}
\arguments{
\item{df}{data frame of a reactive data frame.}

\item{s_fun}{the aggregate function.}
}
\value{
a react table output.
}
\description{
<table> showing the summary of sales, cost & profit from each customer.
}
\examples{
customer_sales_profit_summary(r_df(), "sum")

}
