% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_general.R
\name{date_diff_summary}
\alias{date_diff_summary}
\title{<data frame> with the difference in days between to date variables.}
\usage{
date_diff_summary(
  df,
  date_var1,
  date_var2,
  sumy_var,
  sumy_fun = "mean",
  gp_var,
  multi_fun = TRUE
)
}
\arguments{
\item{df}{data frame.}

\item{date_var1}{first date variable (the most recent of the two date).}

\item{date_var2}{second date variable.}

\item{sumy_var}{the variable to summarise by}

\item{sumy_fun}{the aggregate function.}

\item{gp_var}{if supplied, the variable to group by.}

\item{multi_fun}{whether to return multiple description or just 1 description.}
}
\value{
a tibble.
}
\description{
<data frame> with the difference in days between to date variables.
}
\examples{
date_diff_summary(data, "delivery_date", "order_date", "profit")

}
