% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_component.R
\name{description_block}
\alias{description_block}
\title{<component> description box}
\usage{
description_block(
  textOne = NULL,
  valueOne = NULL,
  textTwo = NULL,
  valueTwo = NULL,
  rightBorder = TRUE,
  textThree = NULL,
  valueThree = NULL,
  iconThree = NULL,
  colorThree = "blue"
)
}
\arguments{
\item{textOne}{header text.}

\item{valueOne}{header value.}

\item{textTwo}{second header text.}

\item{valueTwo}{second header value.}

\item{rightBorder}{whether to add a border at the right side of the box.}

\item{textThree}{third text.}

\item{valueThree}{third value.}

\item{iconThree}{icon to be added to the third value.}

\item{colorThree}{the color of the third value.}
}
\value{
html tool tag.
}
\description{
<component> description box
}
