% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_general.R
\name{dyn_date}
\alias{dyn_date}
\title{<character> returns the appropriate variable based on x}
\usage{
dyn_date(x, recent = TRUE)
}
\arguments{
\item{x}{a variable name.}

\item{recent}{whether it is a recent date variable name or not.}
}
\value{
a character vector.
}
\description{
<character> returns the appropriate variable based on x
}
\examples{
dyn_date("orderdate")

}
