% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_general.R
\name{filter_period}
\alias{filter_period}
\title{<data frame> Filter Date}
\usage{
filter_period(df, y_all = TRUE, y_19 = FALSE, y_20 = FALSE, y_21 = FALSE)
}
\arguments{
\item{df}{a data frame of reactive data frame.}

\item{y_all}{all years}

\item{y_19}{2019}

\item{y_20}{2020}

\item{y_21}{2021}
}
\value{
a subset of the data when a date is true and y_all is false else the
data
}
\description{
<data frame> Filter Date
}
