% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_product.R
\name{ftbl_product_summary_by}
\alias{ftbl_product_summary_by}
\title{<table> showing a variable summary by product.}
\usage{
ftbl_product_summary_by(df, sumy_var)
}
\arguments{
\item{df}{data frame or a reactive data frame.}

\item{sumy_var}{variable to summarise.}
}
\value{
a react table output.
}
\description{
<table> showing a variable summary by product.
}
\examples{
ftbl_product_summary_by(r_df(), "profit")

}
