% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_sales_channel.R
\name{ftbl_sales_channel_var_count}
\alias{ftbl_sales_channel_var_count}
\title{<table> showing the number of transactions involved by each sales channel.}
\usage{
ftbl_sales_channel_var_count(df, variable)
}
\arguments{
\item{df}{a data frame or a reactive data frame.}

\item{variable}{variable to count with sales channel.}
}
\value{
a react table output.
}
\description{
<table> showing the number of transactions involved by each sales channel.
}
\examples{
ftbl_sales_channel_var_count(r_df(), "customer_names")

}
