% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_general.R
\name{get_filter_unique_cat}
\alias{get_filter_unique_cat}
\title{<character> a vector of unique values.}
\usage{
get_filter_unique_cat(f_df, cat_var, sample = TRUE)
}
\arguments{
\item{f_df}{data frame or reactive data frame.}

\item{cat_var}{a character variable.}

\item{sample}{whether to return a sample.}
}
\value{
character vector.
}
\description{
<character> a vector of unique values.
}
\examples{
get_filter_unique_cat(r_df(), "region")

}
