% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_overview.R
\name{get_mom}
\alias{get_mom}
\title{Month on month.}
\usage{
get_mom(df, current_date, a_fun = "sum")
}
\arguments{
\item{df}{data frame or reactive data frame.}

\item{current_date}{current date for the calculation.}

\item{a_fun}{aggregate function.}
}
\value{
numeric value.
}
\description{
Month on month.
}
\examples{
get_mom(r_df, "2019-10-06")

}
