% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_overview.R
\name{get_mtd}
\alias{get_mtd}
\title{Month to date.}
\usage{
get_mtd(df, start_date, end_date, a_fun = "sum")
}
\arguments{
\item{df}{data frame or reactive data frame.}

\item{start_date}{date to begin calculation.}

\item{end_date}{final date of the calculation.}

\item{a_fun}{aggregate function.}
}
\value{
numeric value.
}
\description{
Month to date.
}
