% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_component.R
\name{get_year_choice}
\alias{get_year_choice}
\title{Get the available year choices}
\usage{
get_year_choice(df)
}
\arguments{
\item{df}{data frame or reactive data frame}
}
\value{
a vector
}
\description{
Get the available year choices
}
