% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_overview.R
\name{get_yoy}
\alias{get_yoy}
\title{Year on year}
\usage{
get_yoy(df, period, period_value, a_fun = "sum")
}
\arguments{
\item{df}{data frame or reactive data frame.}

\item{period}{either month or quarter.}

\item{period_value}{which particular month or quarter.}

\item{a_fun}{aggregate function.}
}
\value{
numeric value.
}
\description{
Year on year
}
\examples{
get_yoy(r_df, "month", "March")

}
