% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_component.R
\name{input_switch}
\alias{input_switch}
\title{<component> wrapper for shinyWidget pretty Switch input.}
\usage{
input_switch(id, value = FALSE, in_line = FALSE)
}
\arguments{
\item{id}{input id.}

\item{value}{TRUE or FALSE.}

\item{in_line}{render input inline.}
}
\value{
shiny check box input Widget for showing inputs.
}
\description{
<component> wrapper for shinyWidget pretty Switch input.
}
