% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_component.R
\name{location_picker_input}
\alias{location_picker_input}
\title{<component> wrapper for shinyWidget picker input.}
\usage{
location_picker_input(id, selected = "region", in_line = FALSE)
}
\arguments{
\item{id}{input id.}

\item{selected}{the selected choice.}

\item{in_line}{render input inline.}
}
\value{
shiny select input Widget for geo location choice.
}
\description{
<component> wrapper for shinyWidget picker input.
}
