% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_general.R
\name{location_sub_summary}
\alias{location_sub_summary}
\title{<data frame> summarise a variable geographic location.}
\usage{
location_sub_summary(f_df, gp_vars, sumy_var, s_fun = "sum", group = FALSE)
}
\arguments{
\item{f_df}{data frame or reactive data frame.}

\item{gp_vars}{geo location to group by.}

\item{sumy_var}{the summary variable.}

\item{s_fun}{the aggregate function.}

\item{group}{whether to keep the group index.}
}
\value{
a tibble.
}
\description{
<data frame> summarise a variable geographic location.
}
\examples{
location_sub_summary(r_df(), c("region", "state", "county"), "profit")

}
