% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_product.R
\name{min_max_product_order}
\alias{min_max_product_order}
\title{<table> showing the highest or least order quantity by product and geo location}
\usage{
min_max_product_order(df, gp_var = "region", sumy_fun = "sum", get_max = TRUE)
}
\arguments{
\item{df}{data frame or a reactive data frame.}

\item{gp_var}{variable to group by.}

\item{sumy_fun}{aggregate function.}

\item{get_max}{whether to get the maximum value for each group if true or the minimum}
}
\value{
react table output.
}
\description{
<table> showing the highest or least order quantity by product and geo location
}
\examples{
min_max_product_order(r_df(), gp_var = "region")

}
