% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_overview.R
\name{numeric_lab}
\alias{numeric_lab}
\title{numeric label}
\usage{
numeric_lab(value, output_type = "dollar")
}
\arguments{
\item{value}{numeric value}

\item{output_type}{type of output any of "percent", "dollar" or "number"}
}
\value{
character value.
}
\description{
numeric label
}
\examples{
numeric_lab(2458974, "number")

}
