% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_general.R
\name{numeric_summary}
\alias{numeric_summary}
\title{<data frame> of numeric summary.}
\usage{
numeric_summary(
  df,
  var,
  include_count = TRUE,
  groups = "drop_last",
  pivot = FALSE
)
}
\arguments{
\item{df}{a data frame.}

\item{var}{a numeric variable to summarise.}

\item{include_count}{whether to include the count.}

\item{groups}{how to treat groupings if available.}

\item{pivot}{whether to convert the table to a long format (useful when there are no groupings)}
}
\value{
a tibble.
}
\description{
<data frame> of numeric summary.
}
\examples{
numeric_summary(data, "profit")

}
