% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_component.R
\name{page_analysis_settings}
\alias{page_analysis_settings}
\title{Page analysis filter drop down}
\usage{
page_analysis_settings(
  dropdown_id,
  id_all,
  id_19,
  id_20,
  id_21,
  max_width = 180,
  ...
)
}
\arguments{
\item{dropdown_id}{id of the drop down menu.}

\item{id_all}{id for selecting all the years.}

\item{id_19}{id for selecting 2019.}

\item{id_20}{id for selecting 2020.}

\item{id_21}{id for selecting 2021.}
}
\value{
shiny widgets drop down menu.
}
\description{
Page analysis filter drop down
}
