% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_overview.R
\name{plt_cost_breakdown}
\alias{plt_cost_breakdown}
\title{<plot> of cost break down.}
\usage{
plt_cost_breakdown(df)
}
\arguments{
\item{df}{data frame or a reactive data frame.}
}
\value{
a echarts4r doughnut chart
}
\description{
<plot> of cost break down.
}
\examples{
plt_cost_breakdown(r_df())

}
