% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_customer.R
\name{plt_order_date}
\alias{plt_order_date}
\title{<plot> of variables by order date}
\usage{
plt_order_date(f_df, category, by, cum = FALSE)
}
\arguments{
\item{f_df}{data frame of a reactive data frame.}

\item{category}{the filtered category}

\item{by}{the variable to plot.}

\item{cum}{whether to use a cumulative sum or not.}
}
\value{
a echarts4r line and area chart.
}
\description{
<plot> of variables by order date
}
\examples{
plt_order_date(r_df(), "aaa ltd", "profit")

}
