% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_general.R
\name{plt_percent_profit_cost}
\alias{plt_percent_profit_cost}
\title{<plot> the percentage of profit and cost.}
\usage{
plt_percent_profit_cost(df)
}
\arguments{
\item{df}{data frame of reactive data frame.}
}
\value{
a echarts4r doughnut chart.
}
\description{
<plot> the percentage of profit and cost.
}
\examples{
plt_percent_profit_cost(r_df())

}
