% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_product.R
\name{plt_product_quantity}
\alias{plt_product_quantity}
\title{<plot> number of order for each product.}
\usage{
plt_product_quantity(df, sumy_fun = "sum", top = TRUE, n = 10)
}
\arguments{
\item{df}{data frame or a reactive data frame.}

\item{sumy_fun}{aggregate function.}

\item{top}{whether to count the top n or bottom n.}

\item{n}{number of observations to return.}
}
\value{
a echarts4r bar chart.
}
\description{
<plot> number of order for each product.
}
\examples{
plt_product_quantity(r_df())

}
