% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_overview.R
\name{plt_profit_by_region}
\alias{plt_profit_by_region}
\title{<plot> of profit received in each region.}
\usage{
plt_profit_by_region(df)
}
\arguments{
\item{df}{data frame or a reactive data frame.}
}
\value{
a echarts4r bar chart.
}
\description{
<plot> of profit received in each region.
}
\examples{
plt_profit_by_region(r_df())

}
