% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_overview.R
\name{plt_revenue_year}
\alias{plt_revenue_year}
\title{<plot> of variable summary be year.}
\usage{
plt_revenue_year(df, f_year, sumy_var)
}
\arguments{
\item{df}{data frame or a reactive data frame.}

\item{f_year}{the year to use.}

\item{sumy_var}{variable to summarise.}
}
\value{
a echarts4r line chart
}
\description{
<plot> of variable summary be year.
}
\examples{
plt_revenue_year(r_df(), 2020, "discount")

}
