% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_sales_channel.R
\name{plt_sales_channel_count}
\alias{plt_sales_channel_count}
\title{<plot> number of transaction involved in each sales channel.}
\usage{
plt_sales_channel_count(df)
}
\arguments{
\item{df}{a data frame or a reactive data frame.}
}
\value{
echarts4r doughnut chart.
}
\description{
<plot> number of transaction involved in each sales channel.
}
\examples{
plt_sales_channel_count(r_df())

}
