% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_sales_channel.R
\name{plt_sc_sales_profit}
\alias{plt_sc_sales_profit}
\title{<plot> of sales and profit summary by sales channel.}
\usage{
plt_sc_sales_profit(df, s_fun)
}
\arguments{
\item{df}{a data frame or a reactive data frame.}

\item{s_fun}{aggregate function.}
}
\value{
echarts4r radar chart.
}
\description{
<plot> of sales and profit summary by sales channel.
}
\examples{
plt_sc_sales_profit(r_df(), "mean")

}
