% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_general.R
\name{plt_ser_summary}
\alias{plt_ser_summary}
\title{<plot> of sales, cost & profit by geo location.}
\usage{
plt_ser_summary(f_df, fl_var, category, gp_var, s_fun = "sum", gp_sample)
}
\arguments{
\item{f_df}{a filtered data frame or reactive data frame.}

\item{fl_var}{the filtered variable}

\item{category}{the subset of the variable that was filtered.}

\item{gp_var}{the variable to group by.}

\item{s_fun}{the aggregate function.}

\item{gp_sample}{if the unique values of the grouped variable is greater than 5, then a sample of the values.}
}
\value{
a echarts4r bar chart.
}
\description{
<plot> of sales, cost & profit by geo location.
}
\examples{
plt_ser_summary(r_df(), "customer_name", "Ayomide Samule", "region")

}
