% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_sales_rep.R
\name{plt_sp_region_revenue_summary}
\alias{plt_sp_region_revenue_summary}
\title{<plot> sales, ..  by Sales team for each region.}
\usage{
plt_sp_region_revenue_summary(f_df, variable, s_fun)
}
\arguments{
\item{f_df}{data frame or reactive data frame.}

\item{variable}{the variable to summarise by either sales, cost or profit etc.}

\item{s_fun}{the summary function.}
}
\value{
a echart4r bar chart.
}
\description{
<plot> sales, ..  by Sales team for each region.
}
\examples{
plt_sp_region_revenue_summary(r_df(), "profit", "sum")

}
