% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_sales_rep.R
\name{plt_sp_sales_profit_summary}
\alias{plt_sp_sales_profit_summary}
\title{Plot Sales, Cost & Profit Summary for all sales team}
\usage{
plt_sp_sales_profit_summary(df, use, s_fun)
}
\arguments{
\item{df}{data frame or reactive data frame.}

\item{use}{the variable to summarise by either sales, cost or profit.}

\item{s_fun}{the summary function.}
}
\value{
a echart4r bar chart / stacked bar chart.
}
\description{
Plot Sales, Cost & Profit Summary for all sales team
}
\examples{
plt_sp_sales_profit_summary(r_df(), "profit", "mean")

}
