% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_sales_rep.R
\name{plt_sp_trans_region}
\alias{plt_sp_trans_region}
\title{<Plot> The number of transaction in each region by sales team.}
\usage{
plt_sp_trans_region(f_df)
}
\arguments{
\item{f_df}{data frame or reactive data frame.}
}
\value{
a echart4r pie chart.
}
\description{
<Plot> The number of transaction in each region by sales team.
}
\examples{
plt_sp_trans_region(r_df())

}
