% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_product.R
\name{product_location_qty2}
\alias{product_location_qty2}
\title{<data frame> summarise product and quantity order.}
\usage{
product_location_qty2(df, gp_var, sumy_fun, sort = FALSE, group = FALSE)
}
\arguments{
\item{df}{data frame}

\item{gp_var}{variable to group by.}

\item{sumy_fun}{aggregate function.}

\item{sort}{whether to sort the table in ascending order.}

\item{group}{whether to keep the group index.}
}
\value{
a tibble.
}
\description{
<data frame> summarise product and quantity order.
}
\examples{
product_location_qty2(data, "region", "sum")

}
