% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_overview.R
\name{revenue_summary_list}
\alias{revenue_summary_list}
\title{<list> of total and average summary.}
\usage{
revenue_summary_list(df, sumy_var, desc_pl = 2)
}
\arguments{
\item{df}{data frame or a reactive data frame.}

\item{sumy_var}{variable to summarise.}

\item{desc_pl}{number of decimal points.}
}
\value{
a list of aggregations
}
\description{
<list> of total and average summary.
}
\examples{
revenue_summary_list(r_df(), "profit" 2)

}
