% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_customer.R
\name{rtbl_customer_order_days}
\alias{rtbl_customer_order_days}
\title{<table> showing a description of the number of days since a customer last made a purchase.}
\usage{
rtbl_customer_order_days(f_df, customer)
}
\arguments{
\item{f_df}{data frame of a reactive data frame.}

\item{customer}{the customer name.}
}
\value{
a react table output.
}
\description{
<table> showing a description of the number of days since a customer last made a purchase.
}
\examples{
rtbl_customer_order_days(r_df(), "Ayomide Samuel")

}
