% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_product.R
\name{rtbl_product_datediff}
\alias{rtbl_product_datediff}
\title{<table> showing the difference  between to date in days.}
\usage{
rtbl_product_datediff(df, date_var1, date_var2)
}
\arguments{
\item{df}{data frame or a reactive data frame.}

\item{date_var1}{first date variable (must be the most recent date of the two date variable)}

\item{date_var2}{second date variable.}
}
\value{
a react table output.
}
\description{
<table> showing the difference  between to date in days.
}
\examples{
rtbl_product_datediff(r_df(), "delivery_date", "procurement_date")

}
