% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_product.R
\name{rtbl_product_quantity}
\alias{rtbl_product_quantity}
\title{<table> showing the order summary by product and geo location.}
\usage{
rtbl_product_quantity(df, gp_var2 = "region", sumy_fun)
}
\arguments{
\item{df}{data frame or a reactive data frame.}

\item{gp_var2}{second variable to group by with product.}

\item{sumy_fun}{aggregate function.}
}
\value{
a react table output.
}
\description{
<table> showing the order summary by product and geo location.
}
\examples{
rtbl_product_quantity(r_Df(), "state", "sum")

}
