% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_customer.R
\name{rtbl_sub_cloc_drilldown}
\alias{rtbl_sub_cloc_drilldown}
\title{<table> showing a variable summary by location.}
\usage{
rtbl_sub_cloc_drilldown(f_df, gp_vars, sumy_var, customer_name, s_fun = "sum")
}
\arguments{
\item{f_df}{data frame of a reactive data frame.}

\item{gp_vars}{a variable to group by.}

\item{sumy_var}{a variable to summarise.}

\item{customer_name}{the name of the customer.}

\item{s_fun}{the aggregate function.}
}
\value{
a react table output.
}
\description{
<table> showing a variable summary by location.
}
\examples{
rtbl_sub_cloc_drilldown(r_df(), c("state", "county"), "profit", "Ayomide Samuel")

}
