% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_general.R
\name{rtbl_sub_sale_profit}
\alias{rtbl_sub_sale_profit}
\title{<table> showing the summary based on sales, cost & profit.}
\usage{
rtbl_sub_sale_profit(f_df, count_var, s_fun = "sum", bold_name = FALSE)
}
\arguments{
\item{f_df}{data frame or a reactive data frame.}

\item{count_var}{variable to summarise.}

\item{s_fun}{the aggregate function.}

\item{bold_name}{name the first column name bold.}
}
\value{
react table output.
}
\description{
<table> showing the summary based on sales, cost & profit.
}
\examples{
rtbl_sub_sale_profit(r_Df(), "product_name")

}
