% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_general.R
\name{s_group_num_summary}
\alias{s_group_num_summary}
\title{<data frame> a single description numeric summary.}
\usage{
s_group_num_summary(
  df,
  gp_var1,
  gp_var2,
  sumy_var,
  sumy_fun,
  sort = TRUE,
  group = FALSE
)
}
\arguments{
\item{df}{a data frame.}

\item{gp_var1}{first variable to group by.}

\item{gp_var2}{second variable to group by.}

\item{sumy_var}{the variable to summarise.}

\item{sumy_fun}{the aggregate function.}

\item{sort}{whether to sort the values in ascending format.}

\item{group}{whether to keep the group index.}
}
\value{
a tibble
}
\description{
<data frame> a single description numeric summary.
}
\examples{
s_group_num_summary(data, "region", "sales_channel", "profit", "mean")

}
