% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_component.R
\name{stati_card}
\alias{stati_card}
\title{Statistic card}
\usage{
stati_card(s_id, check_val, value, bg_color = "#FFFFFF", subtitle = " ")
}
\arguments{
\item{s_id}{card id}

\item{check_val}{validate display color and icon.}

\item{value}{display value}

\item{bg_color}{background color.}

\item{subtitle}{card subtitle.}
}
\value{
shiny widgets.
}
\description{
Statistic card
}
