% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_general.R
\name{sub_category_summary}
\alias{sub_category_summary}
\title{<list> of variable summary.}
\usage{
sub_category_summary(filtered_data, full_data = NULL, variable)
}
\arguments{
\item{filtered_data}{filtered data frame or reactive data frame.}

\item{full_data}{the unfiltered data frame or reactive data frame.}

\item{variable}{variable to summarise.}
}
\value{
a list of aggregate summary.
}
\description{
<list> of variable summary.
}
\examples{
sub_category_summary(fr_df(), r_df(), "cost")

}
