% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_general.R
\name{sub_summary}
\alias{sub_summary}
\title{<list> of variable summary.}
\usage{
sub_summary(df, variable, col_name, sum_var, f_data = NULL, currency = FALSE)
}
\arguments{
\item{df}{data frame or reactive data frame.}

\item{variable}{the variable to summarise.}

\item{col_name}{name of the additional summary column.}

\item{sum_var}{the additional summary variable.}

\item{f_data}{if supplied the filtered data frame.}

\item{currency}{whether to format using currency.}
}
\value{
a list of aggregate summary.
}
\description{
<list> of variable summary.
}
