% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_general.R
\name{table_style}
\alias{table_style}
\title{Table theme and language}
\usage{
table_style(type = "theme", cell_padding = 6, info_text = "entries")
}
\arguments{
\item{type}{type of output, either "theme" or "lang".}

\item{cell_padding}{the amount of top, bottom, left & right padding for each cell.}

\item{info_text}{additional info for the table pagination.}
}
\value{
a react table output.
}
\description{
Table theme and language
}
\examples{
reactable(theme = table_style(type = "theme"))

}
