% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_component.R
\name{table_title}
\alias{table_title}
\title{<component> add title to a table.}
\usage{
table_title(title, use_html = FALSE, t_color = "#828282", t_size = 20)
}
\arguments{
\item{title}{title text.}

\item{use_html}{whether to render text as html.}

\item{t_color}{color of the text.}

\item{t_size}{the size of the text.}
}
\value{
html tool h2 tag.
}
\description{
<component> add title to a table.
}
